+++ Location.h
@@ -1,11 +1,3 @@
-
-/* 
- * File:   Location.h
- * Author: NMerkelt
- *
- * Created on 17. Oktober 2018, 13:09
- */
-
 #ifndef LOCATION_H
 #define LOCATION_H
 
@@ -13,18 +5,18 @@
 
 class Location {
 public:
-    Location(float latitude, float longitude);
+    Location(double latitude, double longitude);
     Location(const Location& orig);
     virtual ~Location();
     
-    float getLatitude();
-    float getLongitude();
+    double getLatitude();
+    double getLongitude();
     
     bool isValid();
     QJsonObject toJson();
 private:
-    float latitude;
-    float longitude;
+    double latitude;
+    double longitude;
 };
 
 #endif /* LOCATION_H */
